Create fishvr_utils.py
First utility program. Works with 'move_grating_trackC.py' and with 'grating_track_laser.py. Introduces classes Params (safe+load) and Results (safe+load+plot alphas vs time for left-right and bw-fw as well as fw accelerating, idle grating (for conditioning assay)), tail tracking and choosing of anchor points (3), saving on the fly as .avi file, captions on the video which mark stimuli, display of live view in a named window (cv2) and printing of coords (HERE IT DOES NOT WORK), and finally a test plot to refine the plot output.