Updated imsave2 to imsave2B
Works with fishVR_utils_imsaving2B.py Since the fit queue was getting longer and longer (because tail fitting was not as fast as frame grabbing) the tail tracking was analyzed and improved (profile hooks, line_profiler kernprof). A large amount of time was spent on finding the boarders (to stop fitting)- now calculated by numpy instead of locally in python. The intensity threshhold was changed to a percentile and the step size was changed to 30 pixel (because tail length had an influence on processng time). The selection of anchor points does not show a black image anymore when the camera is not restarted. The two queues save and fit get copies of the frame data so they do not have to share the data.