Create CS_vs_off_all_data
This program reads all spkl files, then sorts the alphas in to CS and off, and takes the absolute values
of all alphas.
mode 1: The upper cut off is 90 (set to 90) and the lower cut off is 5 (set to 0).
The average area is calculated by taking the mean value of all alphas (because the timestamp
is always 1).
mode 2: the binary treshold value is 20 degree. Everything >20 = 1 and <20 = 0. The array is deducted from
the same array shifted to the right by one (to get the derivative).
Then it plots the mean areas/the tail beat count averaged over time span and the SEMs.